requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.12.0
python-multipart==0.0.6
redis==5.0.1
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pydantic>=2.5.0
beautifulsoup4==4.12.2
cloudscraper==1.2.71
libtorrent==2.0.11
aiofiles==23.2.1